board 수정 중간 저장

+++ m-common/src/main/java/com/antmen/antwork/common/api/response/board/BoardListResponseDto.java
@@ -19,5 +19,5 @@ public class BoardListResponseDto {
     private LocalDateTime modifiedAt;
     private Boolean isPinned;
     private Boolean isDeleted;
-    private Integer commentNum;
+    private Long commentNum;
 }
